fix: Move mocks into a testFixture to remove cyclic dependency
