+++ modules/extensions/galasa-extensions-parent/dev.galasa.auth.couchdb/src/test/java/dev/galasa/auth/couchdb/internal/TestCouchdbAuthStore.java
@@ -24,13 +24,13 @@ import dev.galasa.extensions.common.couchdb.pojos.PutPostResponse;
 import dev.galasa.extensions.common.couchdb.pojos.ViewResponse;
 import dev.galasa.extensions.common.couchdb.pojos.ViewRow;
 import dev.galasa.extensions.common.impl.HttpRequestFactoryImpl;
-import dev.galasa.extensions.mocks.BaseHttpInteraction;
-import dev.galasa.extensions.mocks.HttpInteraction;
-import dev.galasa.extensions.mocks.MockCloseableHttpClient;
-import dev.galasa.extensions.mocks.MockHttpClientFactory;
-import dev.galasa.extensions.mocks.MockLogFactory;
-import dev.galasa.extensions.mocks.MockTimeService;
-import dev.galasa.extensions.mocks.couchdb.MockCouchdbValidator;
+import dev.galasa.extensions.common.mocks.BaseHttpInteraction;
+import dev.galasa.extensions.common.mocks.HttpInteraction;
+import dev.galasa.extensions.common.mocks.MockCloseableHttpClient;
+import dev.galasa.extensions.common.mocks.MockHttpClientFactory;
+import dev.galasa.extensions.common.mocks.MockLogFactory;
+import dev.galasa.extensions.common.mocks.MockTimeService;
+import dev.galasa.extensions.common.mocks.couchdb.MockCouchdbValidator;
 import dev.galasa.framework.spi.auth.AuthStoreException;
 import dev.galasa.framework.spi.auth.IFrontEndClient;
 import dev.galasa.framework.spi.auth.IInternalAuthToken;
