+++ modules/extensions/galasa-extensions-parent/dev.galasa.events.kafka/build.gradle
@@ -10,7 +10,7 @@ version = '0.38.0'
 dependencies {
     implementation 'dev.galasa:kafka.clients:3.7.0'
 
-    testImplementation(project(':dev.galasa.extensions.mocks'))
+    testImplementation(testFixtures(project(':dev.galasa.extensions.common')))
 }
 
 // Note: These values are consumed by the parent build process
