+++ modules/extensions/galasa-extensions-parent/dev.galasa.events.kafka/src/test/java/dev/galasa/events/kafka/TestKafkaEventsService.java
@@ -17,9 +17,9 @@ import dev.galasa.events.kafka.internal.KafkaEventsService;
 import dev.galasa.events.kafka.internal.KafkaException;
 import dev.galasa.events.kafka.mocks.MockEventProducer;
 import dev.galasa.events.kafka.mocks.MockEventProducerFactory;
-import dev.galasa.extensions.mocks.MockEnvironment;
-import dev.galasa.extensions.mocks.cps.MockConfigurationPropertyStoreService;
-import dev.galasa.extensions.mocks.events.MockEvent;
+import dev.galasa.extensions.common.mocks.MockEnvironment;
+import dev.galasa.extensions.common.mocks.cps.MockConfigurationPropertyStoreService;
+import dev.galasa.extensions.common.mocks.events.MockEvent;
 import dev.galasa.framework.spi.ConfigurationPropertyStoreException;
 import dev.galasa.framework.spi.EventsException;
 import dev.galasa.framework.spi.events.IEvent;
