+++ modules/extensions/galasa-extensions-parent/dev.galasa.events.kafka/src/test/java/dev/galasa/events/kafka/TestKafkaEventsServiceRegistration.java
@@ -14,7 +14,7 @@ import java.util.List;
 
 import dev.galasa.events.kafka.internal.KafkaEventsService;
 import dev.galasa.events.kafka.internal.KafkaEventsServiceRegistration;
-import dev.galasa.extensions.mocks.MockFrameworkInitialisation;
+import dev.galasa.extensions.common.mocks.MockFrameworkInitialisation;
 import dev.galasa.framework.spi.IEventsService;
 
 public class TestKafkaEventsServiceRegistration {
