+++ modules/extensions/galasa-extensions-parent/dev.galasa.extensions.common/src/testFixtures/java/dev/galasa/extensions/common/mocks/MockStatusLine.java
@@ -3,7 +3,7 @@
  *
  * SPDX-License-Identifier: EPL-2.0
  */
-package dev.galasa.extensions.mocks;
+package dev.galasa.extensions.common.mocks;
 
 import org.apache.http.HttpStatus;
 import org.apache.http.ProtocolVersion;
