+++ modules/extensions/galasa-extensions-parent/dev.galasa.extensions.common/src/testFixtures/java/dev/galasa/extensions/common/mocks/couchdb/MockCouchdbValidator.java
@@ -3,7 +3,7 @@
  *
  * SPDX-License-Identifier: EPL-2.0
  */
-package dev.galasa.extensions.mocks.couchdb;
+package dev.galasa.extensions.common.mocks.couchdb;
 
 import java.net.URI;
