+++ modules/extensions/galasa-extensions-parent/dev.galasa.extensions.common/src/testFixtures/java/dev/galasa/extensions/common/mocks/events/MockEvent.java
@@ -1,4 +1,9 @@
-package dev.galasa.extensions.mocks.events;
+/*
+ * Copyright contributors to the Galasa project
+ *
+ * SPDX-License-Identifier: EPL-2.0
+ */
+package dev.galasa.extensions.common.mocks.events;
 
 import java.util.UUID;
