--- modules/extensions/galasa-extensions-parent/dev.galasa.extensions.mocks/bnd.bnd
@@ -1,19 +1,0 @@
--snapshot ${tstamp}
-Bundle-Name: dev.galasa.extensions.mocks
-Bundle-Description: Galasa RAS - CouchDB
-Bundle-License: https://www.eclipse.org/legal/epl-2.0
-Import-Package: com.google.gson,\
-    dev.galasa,\
-    dev.galasa.framework.spi,\
-    dev.galasa.framework.spi.ras,\
-    dev.galasa.framework.spi.teststructure,\
-    dev.galasa.framework.spi.utils,\
-    javax.validation.constraints;resolution:=optional,\
-    org.apache.commons.logging,org.apache.http,\
-    org.apache.http.client.methods,\
-    org.apache.http.client.utils,\
-    org.apache.http.entity,\
-    org.apache.http.impl.client,\
-    org.apache.http.util
-Export-Package: dev.galasa.extensions.mocks
-
