--- modules/extensions/galasa-extensions-parent/dev.galasa.extensions.mocks/build.gradle
@@ -1,33 +1,0 @@
-plugins {
-    id 'galasa.extensions'
-}
-
-description = 'Galasa extensions mocks used for tests'
-
-version = '0.38.0'
-
-configurations {
-    implementation.transitive = false
-}
-
-dependencies {
-    implementation ('org.apache.httpcomponents:httpclient-osgi:4.5.13')
-    implementation ('org.apache.httpcomponents:httpcore-osgi:4.4.14')
-    implementation ('com.google.code.gson:gson:2.10.1')
-    implementation ('junit:junit:4.13.1')
-    implementation ('org.assertj:assertj-core:3.16.1')
-    implementation (project(':dev.galasa.extensions.common'))
-}
-
-// Note: These values are consumed by the parent build process
-// They indicate which packages of functionality this OSGi bundle should be delivered inside,
-// or referenced from.
-// The settings here are gathered together by the build process to create a release.yaml file
-// which gathers-up all the packaging metadata about all the OSGi bundles in this component.
-ext.projectName=project.name
-ext.includeInOBR          = false
-ext.includeInMVP          = false
-ext.includeInBOM          = false
-ext.includeInIsolated     = false
-ext.includeInCodeCoverage = false
-ext.includeInJavadoc      = false
