--- modules/extensions/galasa-extensions-parent/dev.galasa.extensions.mocks/settings.gradle
@@ -1,1 +1,0 @@
-rootProject.name = "dev.galasa.extensions.mocks"
